Create InsertionSort.java
The index i indicates the "current card" being inserted into  the hand.  At the beginning of each iteration of the for loop, which is indexed by i, the subarray consisting of elements arr[1 .. i-1] constitutes the currently sorted hand, and the remaining subarray arr[i+1 ..n]  corresponds to the pile of cards still on the table. In fact, elements arr[1 .. i-1] are the elements originally in positions  through i-1 but now in sorted order.